refactor(platforms): move Windows color helpers
